Move etcd and worker cloud-config templates into their packages.  Move default cluster.yaml into root.
