+++ core/etcd/config/templates/cloud-config-etcd
