+++ core/nodepool/config/templates/cloud-config-worker
