+++ core/root/config/templates/cluster.yaml
